Move feature flags to model
